refactor eventhub
